src/requirements.txt
# Pillow 5.4.0 and 5.4.1 has a bug when open a PNG file
#Pillow
Pillow==5.3.0
